refactor: Add useToggleMenu hook
This decouples the logic from the DropdownMenu component.
